Adding retrofit implementation  Adding JSON converter factory - Gson. This version get the data response in JSON file.

+++ app/src/main/java/com/yoprogramo/earthquake/model/Metadata.java
@@ -10,7 +10,7 @@ public class Metadata implements Serializable {
 
     @SerializedName("generated")
     @Expose
-    private Integer generated;
+    private Long generated;
     @SerializedName("url")
     @Expose
     private String url;
@@ -27,11 +27,11 @@ public class Metadata implements Serializable {
     @Expose
     private Integer count;
 
-    public Integer getGenerated() {
+    public Long getGenerated() {
         return generated;
     }
 
-    public void setGenerated(Integer generated) {
+    public void setGenerated(Long generated) {
         this.generated = generated;
     }
